add speed

+++ app/src/main/java/com/openapi/ks/myapp/bean/form/CustomeAddForm.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  */
 
 public class CustomeAddForm implements Serializable {
-    public Integer id;
+    public Long id;
     public String title;
     public String url;
     public String schema;
@@ -18,7 +18,7 @@ public class CustomeAddForm implements Serializable {
 
     }
 
-    public CustomeAddForm(Integer id, String title, String url) {
+    public CustomeAddForm(Long id, String title, String url) {
         this.id = id;
         this.title = title;
         this.url = url;
